Actually we're okay with ambiguous types.
I think this basically reverts b6d006...

The first reason is that the types of some of the operators genuinely
are ambiguous.  Take e.g. `con0`:

    con0 :: forall s {t} {z}. R '[s := t] ~<~ z => t -> V0 z

There's no way given the arguments to `con0` for Haskell to figure out
the variable `s`, and so the compiler gets free choice in how it
satisfies the constraint.

The thing is, we know this... the whole point of `con0` is that it's
never used without the visible type application.  So deferring the
ambiguity to the use site is actually the expected behavior.

The second is that it's simply not the case that `x ~<~ z` and `y ~<~ z`
imply any relationship between `x` and `y`. For example, consider the
row

    R ["K" := Int, "J" := Bool]

We expect to be able to prove

    R [] ~<~ R ["K" := Int, "J" := Bool]
    R ["K" := Int] ~<~ R ["K" := Int, "J" := Bool]
    R ["J" := Bool] ~<~ R ["K" := Int, "J" := Bool]
    R ["K" := Int, "J" := Bool] ~<~ R ["K" := Int, "J" := Bool]

but none of the contained rows have anything in common...